# swagger.yml
swagger: '2.0'
info:
  description: This API is describe the basic operation of M and T system
  version: 1.0.0
  title: M & T API
  contact:
    email: lionel12@163.com

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  /projects/{project_id}:
    get:
      tags:
        - Projects
      summary: get a project by id
      operationId: getProjectById
      description: |
        Get a project with id
      produces:
        - application/json
      parameters:
        - name: project_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Success, one single project payload
          schema:
            $ref: '#/definitions/ProjectItem'
        404:
          description: Project not found
        500:
          description: Internal error
    patch:
      tags:
        - Projects
      summary: update a project by id
      operationId: patchProjectById
      description: |
        Update a project with id, including its status
      consumes:
        - application/json
      produces:
        - applicatoin/json
      parameters:
        - name: project_id
          in: path
          type: integer
          required: true
        - name: payload
          in: body
          description: |
            Payload of update a project
          schema:
            $ref: '#/definitions/CreateProjectItem'
      responses:
        200:
          description: Success, project has been updated.
        400:
          description: Bad input parameter
        404:
          description: Project not found
        500:
          description: Internal error
    delete:
      tags:
        - Projects
      summary: Delete a project by id
      operationId: deleteProjectById
      description: |
        Delete a project with id, soft delete WITHOUT GDPR
      produces:
        - applicatoin/json
      parameters:
        - name: project_id
          in: path
          type: integer
          required: true
      responses:
        204:
          description: Success, no content
        404:
          description: Project not found
        500:
          description: Internal error
...
definitions:
...
  CreateProjectItem:
    type: object
    required:
      - name
      - owner
    properties:
      name:
        type: string
        example: Food To Shandong
      ownerId:
        type: integer
        example: 1
      receiverId:
        type: integer
        example: 2
      status:
        type: string
        enum:
          - ACTIVE
          - COMPLETED
          - VOID
          - HOLD
  ProjectItem:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: Food To Shandong
      owner:
        type: string
        example: Aric
      receiver:
        type: string
        example: Lionel
      status:
        type: string
        enum:
          - CREATED
          - PROCESSING
          - COMPLETED
          - VOID
          - HOLD
      hidden:
        type: boolean
        example: false
      createdAt:
        type: string
        format: date-time
        example: '2016-08-29T09:12:33.001Z'
      endedAt:
        type: string
        format: date-time
        example: '2016-08-29T09:12:33.001Z'
...
host: virtserver.swaggerhub.com
basePath: /Lionel1204/MandT/1.0.1
schemes:
  - https